requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
chess==1.9.2
idna==3.3
numpy==1.23.1
pandas==1.4.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
SciPy==1.8.1
six==1.16.0
urllib3==1.26.10
pyyaml==6.0
addict==2.4.0
flask==2.1.3
